Rename nvtcache to redis-sink
Since the implementation does handle NVTs but is not limited to it it's
renamed to redis-sink so that it is clear that this is the sink
implementation for redis.

This implementation is downwards compatible to the openvas -u storage.
