+++ rust/Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["nasl-syntax", "nasl-interpreter", "nasl-cli", "nvtcache", "sink"]
+members = ["nasl-syntax", "nasl-interpreter", "nasl-cli", "sink", "redis-sink" ]
