+++ rust/redis-sink/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nvtcache"
+name = "redis-sink"
 version = "0.1.0"
 edition = "2021"
