+++ rust/redis-sink/src/dberror.rs
