+++ rust/redis-sink/src/lib.rs
@@ -3,7 +3,5 @@ pub mod dberror;
 /// Module to handle Nvt metadata. The Nvt structure is defined here as well
 /// as the methods to set and get the struct members.
 pub mod nvt;
-/// Module include objects and methods to upload an Nvt in redis
-pub mod nvtcache;
 /// Module with structures and methods to access redis.
-pub mod redisconnector;
+pub mod connector;
